MakeDF4Regress Define a function to make two dataframe for regression analysis This function was developed based on HSC_multiome_Het_2.ipynb

MakeDF4Regress Define a function to make two dataframe for regression analysis This function was developed based on HSC_multiome_Het_2.ipynb

MakeDF4Regress(
  multiome_wrapper = Donor04_HSC_Multiome_wrapper,
  redeemR = DN4_stemcell_redeemR.seed.sensitive,
  progeny_np = DN4_HSC_LSI_progeny_np,
  assay = "SCT",
  useNPimputation = T,
  maxcloneUMI = 10
)

Arguments

multiome_wrapper

This outject should includes all and more than HSCs cells in redeemR

redeemR

scredeemR object for HSC

progeny_np

run via ProgenyMapping_np

assay

SCT for expression, ATAC for ATAC

useNPimputation

default is T, use all cells called by network propagation, inaddition to the top cells in redeemR

maxcloneUMI

default is 10, Only include genes, in the max clone the expression greater than 10

Value

list(mtx.clone=mtx.clone,mtx.clone.norm.scale=mtx.clone.norm.scale)