Function to filter variants, v4

This function allows you to filter variants,deprecated, use Vfilter_v4 instead

Vfilter_v4(
  InputSummary = VariantsGTSummary,
  Min_Cells = 2,
  Max_Count_perCell = 2,
  QualifyCellCut = 10
)

Arguments

InputSummary

The GTSummary file read in by function CW_mgatk.read

Min_Cells

Default 2, A qualified variant needs the minimum number of cells that have this variant

Max_Count_perCell

Default 2, A qualified variant needs to show at least 2 counts in one cell

QualifyCellCut

Default 10, Minimum depth for a qualified cell

depth

The .depth file by function DepthSummary

Rmvhomo

Boolean (Default F) If true, remove the homozygous variants

Value

this returns feature.list

Examples

DN1CD34_1.Variants.feature.lst<-Vfilter_v3(InputSummary=DN1CD34_1.VariantsGTSummary,depth=DN1CD34_1.depth)
#> Vfilter_v3 has been deprecated, please use Vfilter_v4
#> Error in Vfilter_v3(InputSummary = DN1CD34_1.VariantsGTSummary, depth = DN1CD34_1.depth): object 'DN1CD34_1.VariantsGTSummary' not found