Function to read in redeemV outputs

This function allows you to read raw data from XX/final folder, the output from redeemV It process the data same way as CW_mgatk.read but need to specify one threadhold(thr)

Usage

redeemR.read(
  path,
  thr = "S",
  Processed = F,
  rdsname = "/VariantsGTSummary.RDS"
)

Arguments

path

The XX/final folder, the output from mitoV

thr

The thredhold of filtering T(Total),LS(Less Stringent:c=0.75,a1=2,a2=1), S(Stringent:c=0.75,a1=3,a2=2), VS(Very Stringent:c=0.75,a1=4,a2=3)"

Processed

Boolean variable (Default F), if true directly readRDS("VariantsGTSummary.RDS") or, generate and saveout "VariantsGTSummary.RDS"

Value

this returns depth which is a list of 4 df (Total/LS/S/VS), each is a genotype summary

Examples

if (FALSE) {
WD <- "/path/to/redeemV/final"
DN1CD34_1.VariantsGTSummary <- redeemR.read(WD, thr = "S")
}