Create_mitoTracing

This function is to create mitoTracing with basic information

Create_mitoTracing(
  GTsummary_list,
  depth_list,
  feature.list_list,
  meta_list,
  labels,
  thr = "VerySensitive",
  qualifiedCellCut = 10,
  OnlyHetero = T,
  VAFcut = 1,
  Cellcut = 2,
  maxctscut = 2
)

Arguments

GTsummary_list

simply put GTSummary (Generated by CW_mgatk.read) into list, this allows mergeing multiple dataset this way.

depth_list

simply put depth(Generated by DepthSummary) into list, this allows mergeing multiple dataset this way.

feature.list_list

simply put feature.list(Generated by Vfilter_v3) into list, this allows mergeing multiple dataset this way.

labels

a vector of labels for the samples.

thr

One of the following "Total","VerySensitive","Sensitive","Specific"

qualifiedCellCut

The minimum median mitochondrial coverage for a qualified cell, default is 10

OnlyHetero

If only consider the heteroplasmy variants, default is T

VAFcut

only use variants with VAF smaller than VAFcut. Default is 1. We can use smaller value to constrain into only using rare variants

Cellcut

only use variants with at least cellcut cells carry

maxctscut

only use variants with at least in one cell with at leaset maxctscut variant fragments

Value

mitoTracing class